requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pandas
numpy
fubon-neo
